frontend/backend/requirements.txt
asgiref==3.7.2
Django==5.0.3
gunicorn==21.2.0
packaging==23.2
sqlparse==0.4.4
psycopg2-binary==2.9.9 
# Core
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9   # PostgreSQL
sqlparse==0.4.4
asgiref==3.7.2

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0        # Static file handling if not using S3
packaging==23.2

# Environment variables
python-dotenv==1.0.1
